Notify systemd when we are ready to start serving.
This makes it possible to configure systemd to start wikijs with
ServiceType=notify. This allows system administrators to use systemd
dependencies to delay starting other things (such as proxies) until
wiki-js is ready to accept requests.

Without this, it is difficult to configure wiki-js to start on demand,
without the first couple of requests returning errors.