feat: Allow listening on domain sockets and fds.
When running wikijs behind a proxy (like nginx), it can be simplier to
have the local traffic be sent over a unix domain socket as
adminstrators do not need to worry about allocating local ports.

We also add support for accepting an already listening socket (either
tcp, or a unix domain). This makes it easier to run wikijs under systemd
with socket activation.