fix: Bump jwa version to fix issue in node 25.
The `buffer-equal-constant-time` library (pulled in by the `jwa`
library) is a polyfill adding support for comparing buffers in node.
This library has not been needed for ~5 years. The underlying
functionality this library used was deprecated in node24 and removed in
node25.

This library was being used due to our dependency on an old version of
`jwa`. We bump this dependency to the current version (which doesn't
_appear_ to break anything...)